Create individual project cards from cards class
