+++ src/Dealer.ts
@@ -65,7 +65,26 @@ import { Thorgate } from "./cards/corporation/Thorgate";
 import { UnitedNationsMarsInitiative } from "./cards/corporation/UnitedNationsMarsInitiative";
 
 // Project Cards
-import * as Cards from "./cards/Cards";
+import { AICentral } from "./cards/AICentral";
+import { AcquiredCompany } from "./cards/AcquiredCompany";
+import { AdaptationTechnology } from "./cards/AdaptationTechnology";
+import { AdaptedLichen } from "./cards/AdaptedLichen";
+import { AdvancedAlloys } from "./cards/AdvancedAlloys";
+import { AdvancedEcosystems } from "./cards/AdvancedEcosystems";
+import { AerobrakedAmmoniaAsteroid } from "./cards/AerobrakedAmmoniaAsteroid";
+import { AntiGravityTechnology } from "./cards/AntiGravityTechnology";
+import { Algae } from "./cards/Algae";
+import { Ants } from "./cards/Ants";
+import { AquiferPumping } from "./cards/AquiferPumping";
+import { ArchaeBacteria } from "./cards/ArchaeBacteria";
+import { ArcticAlgae } from "./cards/ArcticAlgae";
+import { ArtificialLake } from "./cards/ArtificialLake";
+import { ArtificialPhotosynthesis } from "./cards/ArtificialPhotosynthesis";
+import { Asteroid } from "./cards/Asteroid";
+import { AsteroidMining } from "./cards/AsteroidMining";
+import { AsteroidMiningConsortium } from "./cards/AsteroidMiningConsortium";
+import { BeamFromAThoriumAsteroid } from "./cards/BeamFromAThoriumAsteroid";
+import { BigAsteroid } from "./cards/BigAsteroid";
 import { BiomassCombustors } from "./cards/BiomassCombustors";
 import { Birds } from "./cards/Birds";
 import { BlackPolarDust } from "./cards/BlackPolarDust";
@@ -331,26 +350,26 @@ export const ALL_PRELUDE_PROJECTS_CARDS: Array<IProjectCard> = [
 
 
 export const ALL_PROJECT_CARDS: Array<IProjectCard> = [
-    new Cards.AcquiredCompany(),
-    new Cards.AdaptationTechnology(),
-    new Cards.AdaptedLichen(),
-    new Cards.AdvancedAlloys(),
-    new Cards.AdvancedEcosystems(),
-    new Cards.AerobrakedAmmoniaAsteroid(),
-    new Cards.AntiGravityTechnology(),
-    new Cards.Ants(),
-    new Cards.AquiferPumping(),
-    new Cards.AICentral(),
-    new Cards.Algae(),
-    new Cards.ArchaeBacteria(),
-    new Cards.ArcticAlgae(),
-    new Cards.ArtificialLake(),
-    new Cards.ArtificialPhotosynthesis(),
-    new Cards.Asteroid(),
-    new Cards.AsteroidMining(),
-    new Cards.AsteroidMiningConsortium(),
-    new Cards.BeamFromAThoriumAsteroid(),
-    new Cards.BigAsteroid(),
+    new AcquiredCompany(),
+    new AdaptationTechnology(),
+    new AdaptedLichen(),
+    new AdvancedAlloys(),
+    new AdvancedEcosystems(),
+    new AerobrakedAmmoniaAsteroid(),
+    new AntiGravityTechnology(),
+    new Ants(),
+    new AquiferPumping(),
+    new AICentral(),
+    new Algae(),
+    new ArchaeBacteria(),
+    new ArcticAlgae(),
+    new ArtificialLake(),
+    new ArtificialPhotosynthesis(),
+    new Asteroid(),
+    new AsteroidMining(),
+    new AsteroidMiningConsortium(),
+    new BeamFromAThoriumAsteroid(),
+    new BigAsteroid(),
     new BiomassCombustors(),
     new Birds(),
     new BlackPolarDust(),
