+++ tests/cards/AdaptationTechnology.spec.ts
@@ -1,6 +1,6 @@
 
 import { expect } from "chai";
-import { AdaptationTechnology } from "../../src/cards/Cards";
+import { AdaptationTechnology } from "../../src/cards/AdaptationTechnology";
 import { Color } from "../../src/Color";
 import { Player } from "../../src/Player";
 import { Game } from "../../src/Game";
