+++ tests/cards/AdvancedAlloys.spec.ts
@@ -1,6 +1,6 @@
 
 import { expect } from "chai";
-import { AdvancedAlloys } from "../../src/cards/Cards";
+import { AdvancedAlloys } from "../../src/cards/AdvancedAlloys";
 import { Color } from "../../src/Color";
 import { Player } from "../../src/Player";
 import { Game } from "../../src/Game";
