+++ tests/cards/AdvancedEcosystems.spec.ts
@@ -1,6 +1,6 @@
 
 import { expect } from "chai";
-import { AdvancedEcosystems } from "../../src/cards/Cards";
+import { AdvancedEcosystems } from "../../src/cards/AdvancedEcosystems";
 import { Color } from "../../src/Color";
 import { Player } from "../../src/Player";
 import { Game } from "../../src/Game";
