+++ tests/cards/AerobrakedAmmoniaAsteroid.spec.ts
@@ -1,6 +1,7 @@
 
 import { expect } from "chai";
-import { AerobrakedAmmoniaAsteroid, Ants } from "../../src/cards/Cards";
+import { AerobrakedAmmoniaAsteroid } from "../../src/cards/AerobrakedAmmoniaAsteroid";
+import { Ants } from "../../src/cards/Ants";
 import { Color } from "../../src/Color";
 import { Player } from "../../src/Player";
 import { Game } from "../../src/Game";
