+++ tests/cards/Algae.spec.ts
@@ -1,6 +1,6 @@
 
 import { expect } from "chai";
-import { Algae } from "../../src/cards/Cards";
+import { Algae } from "../../src/cards/Algae";
 import { Color } from "../../src/Color";
 import { Player } from "../../src/Player";
 import { Game } from "../../src/Game";
