+++ tests/cards/Ants.spec.ts
@@ -1,6 +1,6 @@
 
 import { expect } from "chai";
-import { Ants } from "../../src/cards/Cards";
+import { Ants } from "../../src/cards/Ants";
 import { Color } from "../../src/Color";
 import { Player } from "../../src/Player";
 import { Game } from "../../src/Game";
