+++ tests/cards/ArchaeBacteria.spec.ts
@@ -1,6 +1,6 @@
 
 import { expect } from "chai";
-import { ArchaeBacteria } from "../../src/cards/Cards";
+import { ArchaeBacteria } from "../../src/cards/ArchaeBacteria";
 import { Color } from "../../src/Color";
 import { Player } from "../../src/Player";
 import { Game } from "../../src/Game";
