+++ tests/cards/OptimalAerobraking.spec.ts
@@ -4,7 +4,7 @@ import { OptimalAerobraking } from "../../src/cards/OptimalAerobraking";
 import { Color } from "../../src/Color";
 import { Player } from "../../src/Player";
 import { Game } from "../../src/Game";
-import { BigAsteroid } from "../../src/cards/Cards";
+import { BigAsteroid } from "../../src/cards/BigAsteroid";
 
 describe("OptimalAerobraking", function () {
     it("Should play", function () {
