+++ tests/cards/SymbioticFungus.spec.ts
@@ -4,7 +4,7 @@ import { SymbioticFungus } from "../../src/cards/SymbioticFungus";
 import { Color } from "../../src/Color";
 import { Player } from "../../src/Player";
 import { Game } from "../../src/Game";
-import { Ants } from "../../src/cards/Cards";
+import { Ants } from "../../src/cards/Ants";
 
 describe("SymbioticFungus", function () {
     it("Can't act or play", function () {
