+++ tsconfig.json
@@ -11,7 +11,7 @@
     "removeComments": true,
     "preserveConstEnums": true,
     "skipLibCheck": true,
-    "sourceMap": false,
+    "sourceMap": true,
     "target": "es5",
     "types": ["node"],
     "moduleResolution": "node"
